organize components for Cities and Settings
