+++ app/src/main/java/ua/sviatkuzbyt/cityweather/data/database/DatabaseRoom.kt
@@ -2,6 +2,7 @@ package ua.sviatkuzbyt.cityweather.data.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import ua.sviatkuzbyt.cityweather.data.database.entities.CityEntity
 
 @Database(
     entities = [CityEntity::class],
