+++ app/src/main/java/ua/sviatkuzbyt/cityweather/data/repositories/ForecastRepository.kt
@@ -2,13 +2,12 @@ package ua.sviatkuzbyt.cityweather.data.repositories
 
 import com.google.gson.Gson
 import ua.sviatkuzbyt.cityweather.data.api.getApiResponse
-import ua.sviatkuzbyt.cityweather.data.settingsstore.SettingsStoreManager
-import ua.sviatkuzbyt.cityweather.data.structures.forecast.ForecastData
-import ua.sviatkuzbyt.cityweather.data.structures.forecast.ForecastFiveDaysItem
-import ua.sviatkuzbyt.cityweather.data.structures.forecast.ForecastFiveDaysResponse
-import ua.sviatkuzbyt.cityweather.data.structures.forecast.ForecastTodayDataResponse
-import ua.sviatkuzbyt.cityweather.data.structures.forecast.ForecastTodayResponseItem
-import ua.sviatkuzbyt.cityweather.data.structures.forecast.NoFormatFiveDaysData
+import ua.sviatkuzbyt.cityweather.data.structures.weather.forecast.ForecastData
+import ua.sviatkuzbyt.cityweather.data.structures.weather.forecast.ForecastFiveDaysItem
+import ua.sviatkuzbyt.cityweather.data.structures.weather.forecast.ForecastFiveDaysResponse
+import ua.sviatkuzbyt.cityweather.data.structures.weather.forecast.ForecastTodayDataResponse
+import ua.sviatkuzbyt.cityweather.data.structures.weather.forecast.ForecastTodayResponseItem
+import ua.sviatkuzbyt.cityweather.data.structures.weather.forecast.NoFormatFiveDaysData
 import ua.sviatkuzbyt.cityweather.data.structures.weather.UnitsData
 import ua.sviatkuzbyt.cityweather.data.structures.weather.WeatherItemAppearance.Companion.getWeatherItemAppearance
 import java.time.Instant
@@ -16,20 +15,18 @@ import java.time.ZoneId
 import java.time.format.DateTimeFormatter
 
 class ForecastRepository(
-    private val settingsStoreManager: SettingsStoreManager,
+    private val settingsRepository: SettingsRepository,
     private val gson: Gson
 ) {
-
-
     suspend fun loadForecastToday(city: String): List<ForecastData> {
-        val units = UnitsData(settingsStoreManager.getUnits())
+        val units = UnitsData(settingsRepository.getSettings(SettingsId.Units))
         val apiResponse = getApiResponse(city, "forecast", "&cnt=8&units=${units.unitsApi}")
         val forecastTodayResponse = gson.fromJson(apiResponse, ForecastTodayDataResponse::class.java)
         return formatDataToday(forecastTodayResponse.list, units.temp, units.wind)
     }
 
     suspend fun loadForecastFiveDays(city: String): List<ForecastData> {
-        val units = UnitsData(settingsStoreManager.getUnits())
+        val units = UnitsData(settingsRepository.getSettings(SettingsId.Units))
         val apiResponse = getApiResponse(city, "forecast", "&units=${units.unitsApi}")
         val forecastFiveDayResponse = gson.fromJson(apiResponse, ForecastFiveDaysResponse::class.java)
         return formatDataFiveDays(forecastFiveDayResponse.list, units.temp)
