+++ app/src/main/java/ua/sviatkuzbyt/cityweather/data/repositories/SettingsRepository.kt
@@ -1,25 +1,56 @@
 package ua.sviatkuzbyt.cityweather.data.repositories
 
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.core.edit
+import androidx.datastore.preferences.core.stringPreferencesKey
+import androidx.datastore.preferences.preferencesDataStore
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.map
 import ua.sviatkuzbyt.cityweather.R
-import ua.sviatkuzbyt.cityweather.data.settingsstore.SettingsStoreManager
 import ua.sviatkuzbyt.cityweather.data.structures.settings.SettingsItemData
 
-private const val UNITS_ID = 1
+private val Context.settingsStore: DataStore<Preferences> by preferencesDataStore(name = "settings")
 
-class SettingsRepository(private val settingsStoreManager: SettingsStoreManager) {
+enum class SettingsId{ Units }
 
-    suspend fun getSettingsList() = listOf(
+class SettingsRepository(private val context: Context) {
+    private val unitsKey = stringPreferencesKey("units")
+
+    suspend fun getSettingsInList() = listOf(
         SettingsItemData(
-            id = UNITS_ID,
+            id = SettingsId.Units,
             label = R.string.units,
-            setValue = settingsStoreManager.getUnits(),
+            setValue = getData(unitsKey, "metric"),
             listOfValues = { listOf("standard", "metric", "imperial") }
         )
     )
 
-    suspend fun setSettings(id: Int, value: String){
+    suspend fun setSettings(id: SettingsId, value: String){
         when(id){
-            UNITS_ID -> settingsStoreManager.setUnits(value)
+            SettingsId.Units -> setData(unitsKey, value)
+        }
+    }
+
+    suspend fun getSettings(id: SettingsId): String {
+        return when(id){
+            SettingsId.Units -> getData(unitsKey, "metric")
+        }
+    }
+
+    private suspend fun <T> getData(
+        key: Preferences.Key<T>,
+        default: T
+    ): T {
+        return context.settingsStore.data
+            .map { it[key] ?: default }
+            .first()
+    }
+
+    private suspend fun <T>setData(key: Preferences.Key<T>, value: T){
+        context.settingsStore.edit {
+            it[key] = value
         }
     }
 }
