+++ app/src/main/java/ua/sviatkuzbyt/cityweather/data/structures/settings/SettingsItemData.kt
@@ -1,9 +1,10 @@
 package ua.sviatkuzbyt.cityweather.data.structures.settings
 
 import androidx.annotation.StringRes
+import ua.sviatkuzbyt.cityweather.data.repositories.SettingsId
 
 data class SettingsItemData(
-    val id: Int,
+    val id: SettingsId,
     @StringRes val label: Int,
     val setValue: String,
     val listOfValues: () -> List<String>,
