+++ app/src/main/java/ua/sviatkuzbyt/cityweather/data/structures/weather/WeatherItemAppearance.kt
@@ -2,7 +2,7 @@ package ua.sviatkuzbyt.cityweather.data.structures.weather
 
 import androidx.annotation.DrawableRes
 import ua.sviatkuzbyt.cityweather.R
-import ua.sviatkuzbyt.cityweather.data.structures.cities.CityBackground
+import ua.sviatkuzbyt.cityweather.data.structures.weather.cities.CityBackground
 
 data class WeatherItemAppearance(
     @DrawableRes val icon: Int,
