+++ app/src/main/java/ua/sviatkuzbyt/cityweather/data/structures/weather/forecast/TodayResponse.kt
@@ -1,7 +1,7 @@
-package ua.sviatkuzbyt.cityweather.data.structures.forecast
+package ua.sviatkuzbyt.cityweather.data.structures.weather.forecast
 
-import ua.sviatkuzbyt.cityweather.data.structures.weather.Weather
-import ua.sviatkuzbyt.cityweather.data.structures.weather.Wind
+import ua.sviatkuzbyt.cityweather.data.structures.weather.cities.Weather
+import ua.sviatkuzbyt.cityweather.data.structures.weather.cities.Wind
 
 data class ForecastTodayDataResponse(
     val list: List<ForecastTodayResponseItem>
