+++ app/src/main/java/ua/sviatkuzbyt/cityweather/ui/screens/cities/CitiesScreen.kt
@@ -14,7 +14,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import org.koin.androidx.compose.koinViewModel
 import ua.sviatkuzbyt.cityweather.R
-import ua.sviatkuzbyt.cityweather.data.structures.cities.mapCityEntityToItem
+import ua.sviatkuzbyt.cityweather.data.structures.weather.cities.CityItemData
 import ua.sviatkuzbyt.cityweather.ui.ForecastFiveDaysRoute
 import ua.sviatkuzbyt.cityweather.ui.ForecastTodayRoute
 import ua.sviatkuzbyt.cityweather.ui.LocalNavController
@@ -70,7 +70,7 @@ fun CitiesScreen(){
                             val isOpen = openCityItem == index
 
                             CityItem(
-                                data = mapCityEntityToItem(city),
+                                data = CityItemData.map(city),
                                 isOpen = isOpen,
                                 onTodayClick = {
                                     navController.navigate(ForecastTodayRoute(city.name))
