+++ app/src/main/java/ua/sviatkuzbyt/cityweather/ui/screens/forecasts/elements/ForecastItem.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.vectorResource
-import ua.sviatkuzbyt.cityweather.data.structures.forecast.ForecastData
+import ua.sviatkuzbyt.cityweather.data.structures.weather.forecast.ForecastData
 import ua.sviatkuzbyt.cityweather.ui.elements.other.ItemList
 import ua.sviatkuzbyt.cityweather.ui.elements.other.TextBasicCenter
 import ua.sviatkuzbyt.cityweather.ui.elements.other.TextHeadCenter
